+++ src/ts/components/core/input/TextInput.tsx
@@ -34,6 +34,7 @@ const TextInput = ({
     n_blur = 0,
     debounce = false,
     autoComplete = "off",
+    inputProps,
     ...others
 }: Props) => {
 
@@ -70,6 +71,7 @@ const TextInput = ({
 
     return (
         <MantineTextInput
+            {...inputProps}
             data-dash-is-loading={getLoadingState(loading_state) || undefined}
             onChange={(ev) => setVal(ev.currentTarget.value)}
             value={val}