+++ src/ts/components/core/input/Textarea.tsx
@@ -26,6 +26,7 @@ const Textarea = ({
     n_blur = 0,
     debounce = false,
     autoComplete = "off",
+    inputProps,
     ...others
 }: Props) => {
 
@@ -61,6 +62,7 @@ const Textarea = ({
 
     return (
         <MantineTextarea
+            {...inputProps}
             data-dash-is-loading={getLoadingState(loading_state) || undefined}
             autoComplete={autoComplete}
             {...others}